services: one file per class; photo get and post methods' logic are placed into service objects (two classes - two files)
